+++ app/Http/Controllers/Backend/RolesController.php
@@ -5,11 +5,21 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 
 class RolesController extends Controller
 {
+    public $user;
+
+    public function __construct()
+    {
+        $this->middleware(function($request,$next){
+            $this->user = Auth::guard('admin')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      *
@@ -17,6 +27,9 @@ class RolesController extends Controller
      */
     public function index()
     {
+        if(is_null($this->user) || !$this->user->can('role.view')){
+            abort(403, 'Sorry !! You are Unauthorized to View any role !');
+        }
        $roles = Role::all();
        return view('backend.pages.roles.index',compact('roles'));
     }
@@ -28,6 +41,9 @@ class RolesController extends Controller
      */
     public function create()
     {
+        if(is_null($this->user) || !$this->user->can('role.create')){
+            abort(403, 'Sorry !! You are Unauthorized to Create any role !');
+        }
        $all_permission = Permission::all();
        $permission_group = User::getpermissiongroups();
        return view('backend.pages.roles.create',compact('all_permission','permission_group'));
@@ -41,6 +57,9 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
+        if(is_null($this->user) || !$this->user->can('role.create')){
+            abort(403, 'Sorry !! You are Unauthorized to Create any role !');
+        }
         $request->validate([
             'name' => 'required|max:100|unique:roles'
         ],[
@@ -78,6 +97,9 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
+        if(is_null($this->user) || !$this->user->can('role.edit')){
+            abort(403, 'Sorry !! You are Unauthorized to edit any role !');
+        }
         $role = Role::findById($id, 'admin');
         $all_permission = Permission::all();
         $permission_group = User::getpermissiongroups();
@@ -93,6 +115,9 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(is_null($this->user) || !$this->user->can('role.edit')){
+            abort(403, 'Sorry !! You are Unauthorized to edit any role !');
+        }
         $request->validate([
             'name' => 'required|max:100|unique:roles,name,' . $id
         ],[
@@ -121,6 +146,9 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
+        if(is_null($this->user) || !$this->user->can('role.delete')){
+            abort(403, 'Sorry !! You are Unauthorized to delete any role !');
+        }
         $role = Role::findById($id, 'admin');
         if (!is_null($role)) {
             $role->delete();
